+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/animal/negocio/AnimalServices.java
@@ -3,8 +3,8 @@ package br.ufrpe.bsi.mpoo.petSpeed.animal.negocio;
 import java.util.List;
 
 import br.ufrpe.bsi.mpoo.petSpeed.animal.dominio.Animal;
-import br.ufrpe.bsi.mpoo.petSpeed.os.dominio.OrdemServico;
 import br.ufrpe.bsi.mpoo.petSpeed.animal.persistencia.AnimalDAO;
+import br.ufrpe.bsi.mpoo.petSpeed.os.dominio.OrdemServico;
 import br.ufrpe.bsi.mpoo.petSpeed.medico.persistencia.MedicoDAO;
 import br.ufrpe.bsi.mpoo.petSpeed.os.persistencia.OrdemServicoDAO;
