+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/cliente/dominio/Cliente.java
@@ -5,6 +5,7 @@ import java.util.List;
 import br.ufrpe.bsi.mpoo.petSpeed.animal.dominio.Animal;
 import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Endereco;
 import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Pessoa;
+import br.ufrpe.bsi.mpoo.petSpeed.usuario.dominio.Usuario;
 
 public class Cliente {
 
@@ -21,6 +22,17 @@ public class Cliente {
 	private Pessoa dadosPessoais;
 
 	private List<Animal> animais;
+
+	private Usuario usuario;
+
+
+	public Usuario getUsuario() {
+		return usuario;
+	}
+
+	public void setUsuario(Usuario usuario) {
+		this.usuario = usuario;
+	}
