+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/os/negocio/OrdemServicoServices.java
@@ -3,9 +3,9 @@ package br.ufrpe.bsi.mpoo.petSpeed.os.negocio;
 
 import java.util.List;
 
-import br.ufrpe.bsi.mpoo.petSpeed.os.dominio.OrdemServico;
 import br.ufrpe.bsi.mpoo.petSpeed.animal.persistencia.AnimalDAO;
 import br.ufrpe.bsi.mpoo.petSpeed.cliente.persistencia.ClienteDAO;
+import br.ufrpe.bsi.mpoo.petSpeed.os.dominio.OrdemServico;
 import br.ufrpe.bsi.mpoo.petSpeed.medico.persistencia.MedicoDAO;
 import br.ufrpe.bsi.mpoo.petSpeed.os.persistencia.OrdemServicoDAO;
